feat: rotas de delete do usuario

+++ 4.sample-flask-auth/app.py
@@ -78,15 +78,16 @@ def get_user(user_id):
 @login_required
 def delete_user(user_id):
     user = User.query.get(user_id)
+    
+    if user.id == current_user.id:
+        return jsonify({ "message": f"Delecao nao permitida!" }), 403
+
     if user:
         db.session.delete(user)
         db.session.commit()
         return jsonify({ "message": f"Usuario {user_id} deletado com sucesso!" })
+    
     return jsonify({ "message": "Usuario nao encontrado" }), 404
-
-@app.route("/hello-world", methods=["GET"])
-def hellow_world():
-    return "Hello World"
 
 if __name__ == '__main__':
     app.run(debug=True)
